rename environment file in mvx plugin
